refactor(quantize): delete dead code kernels (Refs PMAT-018)
Remove 445 lines of dead code from quantize/mod.rs:
- MultiWarpBatchedQ4KGemvKernel: 231 lines (ZERO tests, never used)
- Dp4aSIMDQ4KGemvKernel: 214 lines (superseded by TrueDp4aQ4KGemvKernel V10)

Also fixes type mismatch in Titan Duel test (usize -> u32).

quantize/mod.rs: 4,650 -> 4,205 lines (-9.6%)
All 188 quantize kernel tests pass.

+++ tests/hardware_canary.rs
@@ -335,7 +335,7 @@ fn titan_duel_numerical_parity() {
     // ============================================================
     // GPU: Generate PTX and execute
     // ============================================================
-    let kernel = GemmKernel::naive(N, N, N);
+    let kernel = GemmKernel::naive(N as u32, N as u32, N as u32);
     let ptx = kernel.emit_ptx();
 
     let mut module = CudaModule::from_ptx(&ctx, &ptx).expect("PTX compilation failed");
